+++ src/polyfills/README.md
@@ -1,0 +1,3 @@
+# Overview
+
+This folder contain various scripts that are used for supporting old NodeJS Versions. I decided group it depending of the dependency that uses the polyfill so if a dependency changes or if removed we can check if polyfill is not longer needed.
