+++ src/polyfills/index.d.ts
@@ -1,0 +1,7 @@
+import { FindLastIndex } from "./Inquirer/findLastIndex.d";
+
+declare global {
+  interface Array<T> {
+    findLastIndex?: FindLastIndex<T>;
+  }
+}
