+++ src/polyfills/index.ts
@@ -1,0 +1,1 @@
+import "./Inquirer/findLastIndex";
